requirements.txt
asgiref==3.8.1
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==33.1.0
gunicorn==21.2.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
